+++ src/webview/show_changes/LogOutput.ts
@@ -96,7 +96,7 @@ export class LogOutput {
 
     const fs = require("fs"); 
     
-    const config = AppConfig.get_app_confg();
+    const config = AppConfig.get_app_config();
     let compile_list = fs.readFileSync(path.join(workspaceUri.fsPath, config.general['compile-list']));
     // Converting to JSON 
     compile_list = JSON.parse(compile_list);